stm32mp1: board: lock OTP used for board and MAC address on ST board
Temporary patch (don't upstream it):
When ST boards (MB1263 for EV1, MB1272 for DK2/DK1) are detected
but the used OTP used are not locked, this patch force the lock of
these OTPs: MAC address = OTP 57 and 58
and board identifier =OTP 59.

This patch solve issue for the first ST boards provisioned
but not locked.

Signed-off-by: Patrick Delaunay <patrick.delaunay@st.com>
Change-Id: I76f1d3cd3a3b43a096bdf5f0a8f547c2a722363e
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/u-boot/+/153405
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: CIBUILD <smet-aci-builds@lists.codex.cro.st.com>